+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added heap growth strategies as a fine-tuning option to `spawn_opt/2,4`
 - Added `crypto:crypto_one_time/4,5` on ESP32
 - Improved nif and port support on STM32
+- Added support for `atomvm:posix_clock_settime/2`
 
 ### Fixed